Project/calander/requirements.txt
msal
requests
python-dotenv
beautifulsoup4
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
playwright
